requirements.txt
amqp==2.1.4
anyjson==0.3.3
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.2
coreapi==2.0.9
coverage==4.2
decorator==4.0.10
Django==1.10
django-cors-middleware==1.3.1
django-extensions==1.7.4
django-filter==1.0.1
django-push-notifications==1.4.1
djangorestframework==3.5.3
drfdocs==0.0.11
fake-factory==0.5.7
ipaddress==1.0.16
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
itypes==1.1.0
kombu==4.0.2
Markdown==2.6.6
mixer==5.5.7
MySQL-python==1.2.5
openapi-codec==1.1.7
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.3.1
prompt-toolkit==1.0.7
ptyprocess==0.5.1
py==1.4.31
pyasn1==0.2.2
Pygments==2.1.3
pyinapp==0.1.3
pytest==3.0.2
pytest-cov==2.3.1
pytest-django==3.0.0
python-dateutil==2.5.3
python-http-client==2.2.1
pytz==2016.7
redis==2.10.5
requests==2.11.1
rsa==3.4.2
sendgrid==3.6.3
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
stripe==1.41.0
traitlets==4.2.2
uritemplate==3.0.0
uWSGI==2.0.14
vine==1.1.3
wcwidth==0.1.7
